+++ src/main/java/com/cloudera/stockcalculator/service/CurrencyRateProvider.java
@@ -3,6 +3,7 @@ package com.cloudera.stockcalculator.service;
 import com.cloudera.stockcalculator.persistence.model.CurrencyRate;
 import com.cloudera.stockcalculator.persistence.repository.CurrencyRateRepository;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.Optional;
 
@@ -12,14 +13,14 @@ public interface CurrencyRateProvider {
 
     CurrencyRateRepository getCurrencyRateRepository();
 
-    Float getRate(Currency sourceCurrency, Date date);
+    BigDecimal getRate(Currency sourceCurrency, Date date);
 
     default CurrencyRate getCurrencyRate(Currency sourceCurrency, Date date) {
         Optional<CurrencyRate> byDateAndSource = getCurrencyRateRepository().findByDateAndSource(date, sourceCurrency);
         if (byDateAndSource.isPresent()) {
             return byDateAndSource.get();
         }
-        Float rate = getRate(sourceCurrency, date);
+        BigDecimal rate = getRate(sourceCurrency, date);
         CurrencyRate currencyRate = new CurrencyRate();
         currencyRate.setSource(sourceCurrency);
         currencyRate.setTarget(getTargetCurrency());
